Move calculation methods; partitioned epoch-rewards reorg, 4 of 5 (#529)
* Add calculation sub-submodule

* Move calculation methods into sub-submodule

* Move unit tests into calculation sub-submodule
