+++ runtime/src/bank/partitioned_epoch_rewards/mod.rs
@@ -1,3 +1,4 @@
+mod calculation;
 mod distribution;
 mod sysvar;
